09 - Linked List (Insert at Head)
Given a pointer to the head of a linked list, insert a new node before
the head. The next value in the new node should point to head and the
data value should be replaced with a given value. Return a reference to
the new head of the list. The head pointer given may be null meaning
that the initial list is empty.

Complete the function insertNodeAtHead with the following parameter(s):

- SinglyLinkedListNode llist: a reference to the head of a list
- int data: the value to insert in the field of the new node