08 - Linked List (Insert at Tail)
You are given the pointer to the head node of a linked list and an
integer to add to the list. Create a new node with the given integer.
Insert this node at the tail of the linked list and return the head node
of the linked list formed after inserting this new node. The given head
pointer may be null, meaning that the initial list is empty.

Complete thefunction with the following parameters:

- SinglyLinkedListNode: a reference to the head of a list
- int data: the data value for the node to insert

Returns:

- SinglyLinkedListNode pointer: reference to the head of the modified
linked list